cdk/requirements.txt
aws-cdk.aws-cloudfront==1.68.0
aws-cdk.aws-codebuild==1.68.0
aws-cdk.aws-codecommit==1.68.0
aws-cdk.aws-codepipeline==1.68.0
aws-cdk.aws-codepipeline-actions==1.68.0
aws-cdk.aws-s3==1.68.0
